feat: vector

+++ language/polkavm-move-native/src/guest/imports.rs
@@ -12,5 +12,5 @@ extern "C" {
 
 #[polkavm_derive::polkavm_import]
 extern "C" {
-    pub fn guest_alloc(size: usize, align: usize) -> u32;
+    pub fn guest_alloc(size: u64, align: u64) -> u32;
 }
